feat: add notification when user is not authorized

+++ frontend/src/shared/ui/Like/index.tsx
@@ -5,6 +5,8 @@ import {
     useDeleteFromFavoritesMutation,
     useGetFavoritesQuery,
  } from '../../../app/api/fetchFavorites';
+import { useAppSelector } from '../../../app/store/hooks';
+import { toast } from 'react-toastify';
 
 type LikeProps = {
     id: number;
@@ -20,6 +22,7 @@ export const Like = ({
     const { data } = useGetFavoritesQuery('');
     const [addFavorite] = useAddFavoriteMutation();
     const [deleteFromFavorites] = useDeleteFromFavoritesMutation();
+    const isAuth = useAppSelector((state) => state.setIsAuth.isAuth);
 
     const [isLike, setIsLike] = useState(false);
 
@@ -29,10 +32,17 @@ export const Like = ({
     }, [data, id]);
 
     const addToFavorite = () => {
-        setIsLike(!isLike);
-        !isLike
-            ? addFavorite({ id })
-            : deleteFromFavorites({ id });
+        if (isAuth) {
+            setIsLike(!isLike);
+
+            !isLike
+                ? addFavorite({ id })
+                : deleteFromFavorites({ id });
+        } else {
+            toast.error(
+                'Пожалуйста, авторизуйтесь для возможности добавления объявлений в избранное'
+            );
+        }  
     };
 
     return (
